requirements.txt
discord.py==1.7.3
yt-dlp==2021.12.1
